ml/requirements.txt
ultralytics
google-generativeai
opencv-python
Pillow
roboflow
numpy
